Add support for sha224 and sha384
It is important to know what these two new hashes are in relation to
their bigger brothers: As in "FIPS 180-4, 6. SECURE HASH ALGORITHMS":
> The specification for SHA-224 is identical to SHA-256, except that
different initial hash values are used, and the final hash value is
truncated to 224 bits for SHA-224.

The same applies to SHA384 in relation to SHA512. The initial hash
values are different so that knowledge of one does not reveal
information about the other.

Why do SHA224 and SHA384 exist?
> SHA224 (and analogously for SHA384), was introduced to match the
second of the security strengths {80, 112, 128, 192, 256} defined in the
"NIST Special Publication 800-57". That security strength itself was
kept because it matches the assessed security strength of 3DES. That
motivation is reported in RFC 3874.

What are the benefits?
> SHA224's output is 28 bytes instead of 32 for SHA256, which is useful
on some occasions, including storing a huge number of hashes or mapping to an
Elliptic Curve Group of some particular size.

Documentation:
https://tools.ietf.org/html/rfc3874
https://csrc.nist.gov/publications/detail/sp/800-57-part-1/rev-5/final.
https://crypto.stackexchange.com/questions/15151/sha-224-purpose

------------------------------------------------------------------------
Note, in this line, how the declaration for SHA224 is done:

```
DECL_PBKDF2(sha224,
            SHA256_CBLOCK,
            SHA224_DIGEST_LENGTH,
            SHA256_CTX,
            SHA224_Init,
            SHA224_Update,
            SHA256_Transform,
            SHA224_Final,
            sha256_cpy,
            sha224_extract,
            sha256_xor)
```

This, as we can see, is mixing keywords from 224 and 256 alike. Note
what is SHA224 specific:
* The name, for obvious reason
* The digest length, as well clear
* The Init function, as described in the FIPS snippet, is the one that
    applies the initial hash-specific value
* The Update and Final functions are specific to the underlying
    libcrypto library
* The extract function is the truncating one, if we compare for example,
    sha256_extract, and sha224_extract, the only difference is the
    truncated last value of the given array.