Get rid of a statically known parameter
PBKDF2 allows for the length of the derived key to be chosen, although,
as stated in its RFC8012:5.2, the maximum effective search space for the
derived key may be limited by the structure of the underlying
pseudorandom function.

In the case of SCRAM (and of our current implementation), this
pseudorandom function is fixed to be HMAC with any of the sha-1 or sha-2
hash functions, and because of that, SCRAM in RFC5802:2.2 enforces such
dkLen to be the output length of the HMAC function, that is, the length
of the underlying hash function used.

This mean that the PBKDF2 functions don't need this parameter, and they
also don't need the computations associated with it. This integer is
known at macro expansion time, it is the _hashsz parameter of the macro.
Together with the computations simplification, we can also get rid of
the multi-block derivation, as we'll essentially have a single block
already defined at compile-time.