Move void to returns
Functions that took a pointer to the field they needed to modify, now do
create such value and return them. This is kind of a step away from a
"statement" and towards an "expression".

This also enforces getting rid of some copying, as we're creating in
place the erlang binary term that will hold the output:

Instead of:
* Creating a binary, calling PBKDF2 with a pointer to it, allocating a
    temporary array to write to, and memcpy'ing this array into the given
    binary when back to the hi_nif function
Rather do:
* Make hi_nif return PBKDF2, which in turn returns PBKDF2_F, which in
    place creates a binary term and extracts the algorithm output into
    the array of this erlang binary, and return this term directly.