Make API consistent with OTP's crypto
Instead of taking numbers and comparing them, compare the appropriate
atom instead.

To do this, I follow here a common pattern seen in other nif libraries,
where to achieve fast comparison of atoms, a nif term is stored for
quick comparison, in the module's private data.

This module's private data is created during load, it's the second
parameter to the load callback, which we make point to whatever data
fits the business logic of this module. This is later accessible through
the enif_priv_data function.

It is important to note that this data is global to all usages of this
module, and hence concurrent writes are not thread-safe. For this, note
that the only write is done during loading, which the BEAM does in a
blocking fashion, and all other accesses are read-only, which could be
done by any process calling a function in this module.

Note as well that in case of an upgrade, we choose here to simply
recreate the same private data, and an old module's private data will be
deallocated (enif_free) when unload is called.