Introduce algorithm resources to the module state
Generated at macro-expansion time, on the declaration of a PBKDF2
function, we'll now generate C structures that will contain all fields
that are necessary for the PBKDF2 computation.

During loading of the module, new resource types of each will be
allocated on the private state of the module, so that we can later
create resources of the indicated type. This is the code pattern
indicated by the official erl_nif documentation.

Note how the HMAC_CTX_ROUND struct contains already in advance all the
fields that will be needed: this I didn't know a priori, and this commit
is just being written a posteriori, after having learnt the correct way
to do so. For example after having failed to extract fields away from a
temporary stack, or initially defining the struct with pointers that I
forgot to destroy on the —not needed anymore— destructor of this
resource.