Introduce timeslicing into the algorithm
Here we finally move away from running the entire algorithm on dirty
schedulers, towards timeslicing the job in place, in the regular
schedulers. Note the flag removed in the ErlNifFunc array.

To do so, the logic for PBKDF2 and PBKDF2_F was heavily refactored.
Now, PBKDF2 creates a enif resource where it will persists the state of
the computation, to keep it available for the next scheduled chunk.
PBKDF2 initialises this state, padding it, running HMAC_INIT and
HMAC_UPDATE with the initial hash and salt, and then calls PBKDF2_F.

PBKDF2_F is the one doing all the logic. Instead of taking the state
into separate many parameters as before, it now follows the enif pattern
of `env, argc, argv[]`, where its only argument input in argv is an
erlang term to the resource. PBKDF2_F runs the computation in as many
iterations as described by the macros `TIMESLICE_PERCENTAGE` and
`ITER_PER_SLOTS`, and the remaining iterations pending, re-schedules
itself accordingly, and returns the result as an erlang binary when
done.