Describe the types of the algorithm
The main type is the `fast_scram_state` record, which stores the current
step (hence determining if it is a client or a server), nonce,
challenge, channel binding data, and scram definitions, as described by
their respective record definitions. There's a map for caching and
storing data as well, like cached scram calculations taken from a
database for example.

`nonce` separates the segment for the client and the server, so that
they can be used separately when needed.

A `challenge` is determined by the salt and iteration count

`channel_binding` requires a channel binding type, as described by
https://tools.ietf.org/html/rfc5929, for example `<<"tls-unique">>`, and
the type's associated binary data.

`scram_definitions` are those described by the scram rfc, where all
calculations will be stored.