Add rules
All rules are functions that take a binary input and a state, and return
a new state if parsing of the binary input was correct, with the
corresponding data updated.

Each one of these rules simply pattern-matches the binary to ensure it
is correct, validates the value given, and injects it in the state.

For the case of extensions, as they're not supported in this project at
the moment, and are optional in the SCRAM specification,
they return a skip_rule key.